+++ backend/pipelines/initial_data_pipeline.py
@@ -6,7 +6,7 @@ from embeddings.embeddings_utils import (
 import hand_normalization.src.main as normalization
 from utils.key_enums import PipelineDictKeys as Keys
 from utils.csv_utils import create_region_csvs, add_embedding_dict_to_csv
-from ann.milvus import add_embeddings_to_milvus
+from vectordb.milvus import add_embeddings_to_milvus
 
 
 def run_initial_data_pipeline(
@@ -39,16 +39,12 @@ def run_initial_data_pipeline(
     """
     ######## STEP 1: Hand normalization #################################
     if normalize:
-        print(
-            "--------------- Hand-Normalization: Load dataset --------------------------------"
-        )
+        print("--------------- Hand-Normalization: Load dataset --------------------------------")
         # dataloader
 
         dataset_base = ImagePathDataset(base_dataset_path)
 
-        print(
-            "--------------- Hand-Normalization: Normalize Images --------------------------------"
-        )
+        print("--------------- Hand-Normalization: Normalize Images --------------------------------")
         # normalize images
 
         for path_dict in dataset_base:
@@ -59,9 +55,7 @@ def run_initial_data_pipeline(
             # save normalized images (path: UUID_HandRegion)
             # for tests and debugs possible to comment this out
             if save_images:
-                normalization.save_region_images(
-                    uuid, regions_dict, region_dataset_path
-                )
+                normalization.save_region_images(uuid, regions_dict, region_dataset_path)
 
     ######## STEP 2: Embeddings #################################
     # relevant für Test-Szenarien
@@ -72,9 +66,7 @@ def run_initial_data_pipeline(
     dataset = DatasetRegionClusters(region_dataset_path)
     # dataset is a list
 
-    print(
-        "--------------- Embeddings: Calculate embeddings --------------------------------"
-    )
+    print("--------------- Embeddings: Calculate embeddings --------------------------------")
     # für Unit_test
     embeddings_all_test = []
